requirements.txt
streamlit>=1.28.0
requests>=2.31.0
pandas>=2.0.0
# NumPy 2.x breaks wheels built for 1.x (common with conda + older PyTorch).
numpy>=1.24.0,<2.0
nltk>=3.8.0
gensim>=4.3.0
scikit-learn>=1.3.0
plotly>=5.18.0
networkx>=3.2.0
biopython>=1.81
chromadb>=0.4.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0